Update RacunSeeder.java
seedovanje fixovano

+++ banka-service/src/main/java/rs/edu/raf/racun/seeder/RacunSeeder.java
@@ -27,6 +27,7 @@ public class RacunSeeder implements CommandLineRunner {
 
     @Autowired
     private DataSource dataSource;
+    boolean reseed = true;
 
     private final ZemljaRepository zemljaRepository;
     private final ValuteRepository valuteRepository;
@@ -69,6 +70,8 @@ public class RacunSeeder implements CommandLineRunner {
     public void run(String... args) throws Exception {
 
         try {
+            if(!reseed)
+                return;
 /*
             uplataRepository.deleteAll();
             uplataRepository.findAll().forEach(System.out::println);
@@ -117,9 +120,7 @@ public class RacunSeeder implements CommandLineRunner {
             if(valuteRepository.findAll().isEmpty())
                 valuteRepository.saveAll(valute);
 
-            MarzniRacun marzniRacun = new MarzniRacun(-1L, -1L, 444000000000000022L, "RSD", "STOCKS", new BigDecimal(10000000), new BigDecimal(10000000), new BigDecimal(500), new BigDecimal(50), false, null);
-            if(marzniRacunRepository.findAll().isEmpty())
-                marzniRacunRepository.save(marzniRacun);
+            
 
             List<Firma> firme = new ArrayList<>();
             Firma f1 = new Firma(-1L, "Nasa banka", "444000000000000022,444000000000000122,444000000000000222"
@@ -142,44 +143,40 @@ public class RacunSeeder implements CommandLineRunner {
                     , new BigDecimal("0"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v2.getOznaka()
                     , true, new BigDecimal("1"), new BigDecimal(100 * 2));
-            if(!(devizniRacunRepository.findByBrojRacunaAndAktivanIsTrue(dr1.getBrojRacuna()).isPresent()))
-                dRacuni.add(dr1);
+            dRacuni.add(dr1);
 
             DevizniRacun dr2 = new DevizniRacun(444000000000000111L, 2L, new BigDecimal("10000")
                     , new BigDecimal("0"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L,v2.getOznaka()
                     , true, new BigDecimal("1"), new BigDecimal(100 * 2));
-            if(!(devizniRacunRepository.findByBrojRacunaAndAktivanIsTrue(dr2.getBrojRacuna()).isPresent()))
-                dRacuni.add(dr2);
+            dRacuni.add(dr2);
 
             DevizniRacun dr3 = new DevizniRacun(444000000000000211L, 1L, new BigDecimal("10000")
                     , new BigDecimal("0"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v3.getOznaka()
                     , true, new BigDecimal("1"), new BigDecimal(100 * 3));
-            if(!(devizniRacunRepository.findByBrojRacunaAndAktivanIsTrue(dr3.getBrojRacuna()).isPresent()))
-                dRacuni.add(dr3);
+            dRacuni.add(dr3);
 
+            devizniRacunRepository.deleteAll();
             devizniRacunRepository.saveAll(dRacuni);
 
             List<PravniRacun> pRacuni = new ArrayList<>();
             PravniRacun pr1 = new PravniRacun(444000000000000022L, -1L, new BigDecimal("10000")
                     , new BigDecimal("10000"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v8.getOznaka(), true);
-            if(!(pravniRacunRepository.findByBrojRacunaAndAktivanIsTrue(pr1.getBrojRacuna()).isPresent()))
-                pRacuni.add(pr1);
+            pRacuni.add(pr1);
 
             PravniRacun pr2 = new PravniRacun(444000000000000122L, -1L, new BigDecimal("10000")
                     , new BigDecimal("10000"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v2.getOznaka(), true);
-            if(!(pravniRacunRepository.findByBrojRacunaAndAktivanIsTrue(pr2.getBrojRacuna()).isPresent()))
-                pRacuni.add(pr2);
+            pRacuni.add(pr2);
 
             PravniRacun pr3 = new PravniRacun(444000000000000222L, -1L, new BigDecimal("10000")
                     , new BigDecimal("10000"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v3.getOznaka(), true);
-            if(!(pravniRacunRepository.findByBrojRacunaAndAktivanIsTrue(pr3.getBrojRacuna()).isPresent()))
-                pRacuni.add(pr3);
+            pRacuni.add(pr3);
 
+            pravniRacunRepository.deleteAll();
             pravniRacunRepository.saveAll(pRacuni);
 
             List<TekuciRacun> tRacuni = new ArrayList<>();
@@ -187,26 +184,26 @@ public class RacunSeeder implements CommandLineRunner {
                     , new BigDecimal("10000"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v8.getOznaka()
                     , true, "Studentski", new BigDecimal("0.5"), new BigDecimal("0"));
-            if(!(tekuciRacunRepository.findByBrojRacunaAndAktivanIsTrue(tr1.getBrojRacuna()).isPresent()))
-                tRacuni.add(tr1);
+            tRacuni.add(tr1);
 
             TekuciRacun tr2 = new TekuciRacun(444000000910000033L, 2L, new BigDecimal("11000")
                     , new BigDecimal("11000"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v8.getOznaka()
                     , true, "Studentski", new BigDecimal("0"), new BigDecimal("300"));
-            if(!(tekuciRacunRepository.findByBrojRacunaAndAktivanIsTrue(tr2.getBrojRacuna()).isPresent()))
-                tRacuni.add(tr2);
+            tRacuni.add(tr2);
 
             TekuciRacun tr3 = new TekuciRacun(444000000920000033L, 1L, new BigDecimal("1000")
                     , new BigDecimal("1000"), 22222L, System.currentTimeMillis()
                     , System.currentTimeMillis() + 5*365*24*60*60*1000L, v8.getOznaka()
                     , true, "Studentski", new BigDecimal("0"), new BigDecimal("200"));
-            if(!(tekuciRacunRepository.findByBrojRacunaAndAktivanIsTrue(tr3.getBrojRacuna()).isPresent()))
-                tRacuni.add(tr3);
+            tRacuni.add(tr3);
 
+            tekuciRacunRepository.deleteAll();
             tekuciRacunRepository.saveAll(tRacuni);
 
-
+            MarzniRacun marzniRacun = new MarzniRacun(-1L, -1L, 444000000000000022L, "RSD", "STOCKS", new BigDecimal(10000000), new BigDecimal(10000000), new BigDecimal(500), new BigDecimal(50), false, null);
+            if(marzniRacunRepository.findAll().isEmpty())
+                marzniRacunRepository.save(marzniRacun);
 
         }catch (Exception e){
             e.printStackTrace();
